requirements.txt
fastapi
uvicorn[standard]
psycopg2-binary
redis
qrcode[pil]
Pillow
requests
python-multipart
